feat: implement deleting the order by user

+++ api/src/controllers/orderController.js
@@ -167,6 +167,10 @@ const deleteOrder = async (req, res) => {
   try {
     const order = await orderService.getOrderById(id);
 
+    if (order.createdBy != user.id && !user.roles.includes(ROLE_ADMIN)) {
+      return res.status(403).send("Access denied");
+    }
+
     await orderService.deleteOrder(id);
 
     res.send("Order deleted successfully.");
